Remove ComposerHelperViews.swift file
