+++ Sources/StreamChatSwiftUI/ChatComposer/CameraPickerDisplayView.swift
@@ -1,0 +1,26 @@
+//
+// Copyright © 2026 Stream.io Inc. All rights reserved.
+//
+
+import StreamChat
+import SwiftUI
+
+/// View displayed when the camera picker is shown.
+struct CameraPickerDisplayView: View {
+    @Binding var selectedPickerState: AttachmentPickerState
+    @Binding var cameraPickerShown: Bool
+
+    var cameraImageAdded: (AddedAsset) -> Void
+
+    var body: some View {
+        Spacer()
+            .fullScreenCover(isPresented: $cameraPickerShown, onDismiss: {
+                selectedPickerState = .photos
+            }) {
+                ImagePickerView(sourceType: .camera) { addedImage in
+                    cameraImageAdded(addedImage)
+                }
+                .edgesIgnoringSafeArea(.all)
+            }
+    }
+}
