--- Sources/StreamChatSwiftUI/ChatComposer/ComposerHelperViews.swift
@@ -1,114 +1,0 @@
-//
-// Copyright © 2026 Stream.io Inc. All rights reserved.
-//
-
-import StreamChat
-import SwiftUI
-
-/// View used to indicate that an asset is a video.
-public struct VideoIndicatorView: View {
-    @Injected(\.images) private var images
-
-    public init() {}
-
-    public var body: some View {
-        BottomLeftView {
-            Image(uiImage: images.videoIndicator)
-                .customizable()
-                .frame(width: 22)
-                .padding(2)
-                .applyDefaultIconOverlayStyle()
-                .modifier(ShadowModifier())
-        }
-    }
-}
-
-/// View displaying the duration of the video.
-public struct VideoDurationIndicatorView: View {
-    @Injected(\.colors) private var colors
-    @Injected(\.fonts) private var fonts
-
-    var duration: String
-
-    public init(duration: String) {
-        self.duration = duration
-    }
-
-    public var body: some View {
-        BottomRightView {
-            Text(duration)
-                .foregroundColor(Color(colors.staticColorText))
-                .font(fonts.footnoteBold)
-                .padding(.all, 4)
-                .modifier(ShadowModifier())
-        }
-    }
-}
-
-/// View shown after the native file picker is closed.
-struct FilePickerDisplayView: View {
-    @Injected(\.fonts) private var fonts
-    @Injected(\.colors) private var colors
-
-    @Binding var filePickerShown: Bool
-    var onFilesPicked: ([URL]) -> Void
-
-    var body: some View {
-        ZStack {
-            Button {
-                filePickerShown = true
-            } label: {
-                Text(L10n.Composer.Files.addMore)
-                    .font(fonts.bodyBold)
-            }
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .foregroundColor(Color(colors.highlightedAccentBackground))
-        .sheet(isPresented: $filePickerShown) {
-            FilePickerView(onFilesPicked: onFilesPicked)
-        }
-    }
-}
-
-/// View displayed when the camera picker is shown.
-struct CameraPickerDisplayView: View {
-    @Binding var selectedPickerState: AttachmentPickerState
-    @Binding var cameraPickerShown: Bool
-
-    var cameraImageAdded: (AddedAsset) -> Void
-
-    var body: some View {
-        Spacer()
-            .fullScreenCover(isPresented: $cameraPickerShown, onDismiss: {
-                selectedPickerState = .photos
-            }) {
-                ImagePickerView(sourceType: .camera) { addedImage in
-                    cameraImageAdded(addedImage)
-                }
-                .edgesIgnoringSafeArea(.all)
-            }
-    }
-}
-
-/// View displayed when there's no access permission to the photo library.
-struct AssetsAccessPermissionView: View {
-    @Injected(\.images) private var images
-
-    var body: some View {
-        AttachmentAccessPermissionView(
-            image: Image(uiImage: images.attachmentPickerPhotosIcon),
-            text: L10n.Composer.Images.noAccessLibrary,
-            onTap: {
-                openSettings()
-            }
-        )
-    }
-
-    private func openSettings() {
-        guard let url = URL(string: UIApplication.openSettingsURLString),
-              UIApplication.shared.canOpenURL(url) else {
-            return
-        }
-        UIApplication.shared.open(url, options: [:], completionHandler: nil)
-    }
-}
